Enable seo friendly URLs
Mechanism inspired by how Next.js works. The route is not determined by
the browser URL but the state that goes with it. That's what `as` stand for.
The only thing to remember here is that a matchUrl graphql query must be
run on the first page load to know what is the initial `as`.

Additionally, since the folks using pigment should not worry about the Router
in Pigment.js, the graphql part is also abstracted away. The only thing needed
is to implement a `matchUrl` function inside of your loader.
(See `demo/src/server/modules/blog/loader.js` to see how it works)