Enable usage of GET params in url
Those will be fused with the url params.
Ex: `/post/1?from=12` gives : `{id: "1", from: "12"}`

With it, the hashHistory needed previously in the styleguide is no
longer useful. Thus, an additional middleware has been added that catches
all the `/_pigment/styleguide/*` to render the styleguide.
For now, there is no SSR for the Styleguide. But it might be useful in the
future since that would mean you could dev in real conditions within the
styleguide.